+++ src/components/SettingsHighlightColorMenu.tsx
@@ -5,7 +5,8 @@ import styled from 'styled-components'
 
 import HighlightColorMenuItem from 'Components/HighlightColorMenuItem'
 import { HighlightColors } from 'Libs/Highlight'
-import { ifProp } from 'Utils/styled'
+import { withSCProps } from 'Utils/react'
+import { color, ifProp } from 'Utils/styled'
 
 /**
  * ColorMenuButton component.
@@ -13,11 +14,36 @@ import { ifProp } from 'Utils/styled'
 const ColorMenuButton = styled(Button)`
   margin-left: 10px;
   margin-top: ${ifProp('minimal', '-20px', 'initial')};
-  width: 90px;
+  width: 109px;
 
   &.pt-button {
     justify-content: flex-end;
+
+    & .pt-button-text {
+      align-items: center;
+      display: inline-flex;
+      flex: 1;
+    }
   }
+`
+
+/**
+ * ColorPreview component.
+ */
+const ColorPreview = withSCProps<ColorPreviewProps, HTMLDivElement>(styled.div)`
+  background-color: ${(props) => color(`log.highlight.${props.highlightColor}.background`)};
+  border-radius: 2px;
+  height: 11px;
+  margin-right: 10px;
+  width: 11px;
+`
+
+/**
+ * ColorName component.
+ */
+const ColorName = styled.div`
+  flex: 1;
+  text-align: right;
 `
 
 /**
@@ -29,20 +55,23 @@ export default class SettingsHighlightColorMenu extends React.Component<Props> {
    * @return Element to render.
    */
   public render() {
-    const { color, onSelect, small } = this.props
+    const { color: highlightColor, onSelect, small } = this.props
 
     return (
       <Popover
         content={
           <Menu>
-            {_.map(HighlightColors, (name) => <HighlightColorMenuItem key={name} color={name} onClick={onSelect} />)}
+            {_.map(HighlightColors, (name) => (
+              <HighlightColorMenuItem key={name} color={name} onClick={onSelect} selected={name === highlightColor} />
+            ))}
           </Menu>
         }
         position={small ? undefined : Position.BOTTOM_RIGHT}
         usePortal={false}
       >
         <ColorMenuButton rightIcon="caret-down" minimal={small || false}>
-          {color}
+          <ColorPreview highlightColor={highlightColor} />
+          <ColorName>{highlightColor}</ColorName>
         </ColorMenuButton>
       </Popover>
     )
@@ -56,4 +85,11 @@ type Props = {
   color: HighlightColors
   onSelect: (color: HighlightColors) => void
   small?: boolean
+}
+
+/**
+ * React Props.
+ */
+type ColorPreviewProps = {
+  highlightColor: HighlightColors
 }